// src/modules/team/constants/team-permission.enum.ts
export enum TeamPermission {
  TEAM_INVITE_MEMBER = 'TEAM_INVITE_MEMBER',
  TEAM_UPDATE_ROLE = 'TEAM_UPDATE_ROLE',
  TEAM_SETTINGS_UPDATE = 'TEAM_SETTINGS_UPDATE',
  PROJECT_CREATE_DELETE = 'PROJECT_CREATE_DELETE',
  PROJECT_INVITE_MEMBER = 'PROJECT_INVITE_MEMBER',
  PROJECT_UPDATE_ROLE = 'PROJECT_UPDATE_ROLE',
}